Fix warning in previous commit. DONT_BUILD.
git-svn-id: https://svn.eduke32.com/eduke32@4929 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/jaudiolib/src/xa.c
@@ -296,7 +296,7 @@ void MV_SetXAPosition(VoiceNode *voice, int32_t position)
 {
     xa_data * xad = (xa_data *) voice->extra;
 
-    if (position < XA_DATA_START || position >= xad->length)
+    if (position < XA_DATA_START || (size_t)position >= xad->length)
         position = XA_DATA_START;
 
     xad->pos = position;
